Support themodernproper.com

+++ crates/schema-org/src/web_page.rs
@@ -8,7 +8,7 @@ use crate::field::{
     WebPageMaintainerFieldEnum, WebPagePublishingPrinciplesFieldEnum, WebPageReviewedByFieldEnum,
     WebPageSubjectOfFieldEnum, WebPageTranslatorFieldEnum, WebPageVideoFieldEnum,
 };
-use crate::helpers::one_or_many;
+use crate::helpers::{one_or_many, one_or_many_string_or_num};
 use crate::{
     Comment, CreativeWork, ImageObject, InteractionCounter, MediaObject, Thing, WebPageElement,
 };
@@ -173,9 +173,9 @@ pub struct WebPage {
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub thumbnail_url: Vec<String>,
     ///<https://schema.org/copyrightYear>
-    #[serde(default, deserialize_with = "one_or_many")]
+    #[serde(default, deserialize_with = "one_or_many_string_or_num")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub copyright_year: Vec<f32>,
+    pub copyright_year: Vec<i32>,
     ///<https://schema.org/citation>
     #[serde(default, deserialize_with = "one_or_many")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
